infrastructure for frontend
a separate cdk stack for the frontend. It contains an s3 bucket, cloudfront distribution, acm cert, and a route53 A record.

The result is that https://myleg.org now points to the content in the s3 bucket. I over-commented so that I never have when I need to do this again it is painfully easy.

Note: the hosted zone was not provisioned via IAC, it was created by AWS when the domain was purchased in Route53. Therefore we pass in it's ID as a parameter instead of creating it.