deployment scripts
deployment scripts to help enable our cicd. Notice the dependency the frontend content deploys have on the frontend infrastructure deploys.

This dependency using outputs.json is used to pass necessary params like the bucket name and cloudfront distribution ID so that they can be updated when the frontend content is deployed.